requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
geoalchemy2==0.15.2
alembic==1.13.3
requests==2.32.3
apscheduler==3.10.4
python-dotenv==1.0.1
mangum==0.17.0
shapely==2.0.6
